% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segqcc-package.R
\docType{package}
\name{segqcc-package}
\alias{segqcc}
\alias{segqcc-package}
\title{segqcc: Change-Point Segmented Control Charts}
\description{
Detects change points in a control-chart statistic (via the 'changepoint'
package) and rebuilds the chart with the centre and the control limits
recomputed within each homogeneous segment. Trending or regime-switching
series are thereby assessed against locally appropriate limits instead of a
single global set.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chrglez/segqcc}
  \item Report bugs at \url{https://github.com/chrglez/segqcc/issues}
}
}
\author{
\strong{Maintainer}: Christian González-Martel
\email{christian.gonzalez@ulpgc.es}

Authors:
\itemize{
  \item Jaime Pinilla Domínguez \email{jaime.pinilla@ulpgc.es}
}
}
\keyword{internal}
