% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.segmented_qcc}
\alias{print.segmented_qcc}
\title{Print a segmented control chart}
\usage{
\method{print}{segmented_qcc}(x, digits = 4, ...)
}
\arguments{
\item{x}{a \code{segmented_qcc} object.}

\item{digits}{number of significant digits for the segment table.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a segmented control chart
}
\examples{
data(segc)
segmented_qcc(segc$value, type = "c")

}
